+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change log
 
-### [Unreleased][unreleased]
+## [Unreleased][unreleased]
 
-### 0.1.0 - 2016-07-05
+## 0.2.0 - 2016-07-18
+
+### Changed
+- RouteResolver is not mandatory in Api class constructor (BaseRouteResolver will be used if is not set)
+- $rules are optional for Validation
+
+### Added
+- getValue($key, $default = null) function to Validation
+
+## 0.1.0 - 2016-07-05
 - First tagged version
 
 [unreleased]: https://github.com/ricco24/api-nette/compare/0.1.0...HEAD